Plot Type 1

Plot type 1: boxes with mean flow as height and event duration as width

hsTimeBarPlot(
  t1,
  t2,
  height,
  tlim = c(min(t1), max(t2)),
  ylim = NULL,
  ylog = TRUE,
  ymult = c(1, 2, 3, 5),
  main = "hsTimeBarPlot",
  tlab = "time",
  ylab = "height",
  time.format = .defaultDateFormat(),
  tlab.mindist = 86400,
  col = "grey",
  cex.all = 0.8,
  cex.legend = cex.all,
  cex.text = cex.all,
  cex.axis = cex.all,
  legend.values = height,
  legend.sort = TRUE,
  legend.format = .defaultLegendFormat(),
  legend.title = "Legend:",
  mar = c(6, 5, 5, 10)
)

Arguments

t1

begin times of bars

t2

end times of bars

height

bar heights

tlim

limits of time axis. Default: c(min(t1), max(t2))

ylim

limits of y axis, Default: c(min(height), max(height))

ylog

if TRUE, y axis is scaled logarithmically, else linearly

ymult

numbers to be multiplied by exponents of ten for labelling of y axis when ylog is TRUE

main

plot title

tlab

label of time axis

ylab

label of y axis

time.format

format of time axis labels

tlab.mindist

minimum "time distance" between time labels, in seconds

col

bar colour

cex.all

general expansion factor to be applied to cex.legend, cex.text, cex.axis, if not given

cex.legend

expansion factor for legend texts

cex.text

expansion factor of texts within plot (bar numbers)

cex.axis

expansion factor of axis labels

legend.values

values to be shown in legend, default: height values

legend.sort

if TRUE, legend is sorted decreasingly by legend.values

legend.format

format string to be used by sprintf for generation of legend entries from legend values, e.g. "V = %8.0f m3"

legend.title

legend title

mar

plot margins